# internal/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Todo struct {
	ID          int64            `json:"id"`
	Title       string           `json:"title"`
	Description pgtype.Text      `json:"description"`
	Completed   bool             `json:"completed"`
	UserID      int64            `json:"user_id"`
	CreatedAt   pgtype.Timestamp `json:"created_at"`
	UpdatedAt   pgtype.Timestamp `json:"updated_at"`
}

type User struct {
	ID        int64            `json:"id"`
	Username  string           `json:"username"`
	Password  string           `json:"password"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
	UpdatedAt pgtype.Timestamp `json:"updated_at"`
}